// src/app/shared/models/parametro.ts
  export interface Parametro {

        idParametro?: number;
        nombreParametro?: string;
        indicadorBaja?: string;
        deValor1?: string;
        deValor2?: string;
        deValor3?: string;
        deParametro?: string;
        coParametro?: string;
        fechaCreacion?: string;
        usuarioModificacion?: string;
        ipModificacion?: string;
        fechaModificacion?: string;
        usuarioCreacion?: string;
        ipCreacion?: string;
        
    }